requirements.txt
smolagents[litellm]
python-dotenv
ddgs
plotly
geopandas
shapely
kaleido
llama-index
llama-index-llms-huggingface
llama-index-llms-huggingface-api
chromadb
llama-index-embeddings-huggingface
llama-index-vector-stores-postgres
llama-index-vector-stores-chroma
langgraph
IPython
langfuse
langchain_ollama
langchain